refactor: remove deprecated authentication route files for users and modularize route management, improving clarity and maintainability
