--- backend/routes/auth/routes/index.js
@@ -1,9 +1,0 @@
-import express from 'express';
-import userRoutes from './userRoutes.js';
-
-const router = express.Router();
-
-// Mount route modules
-router.use('/users', userRoutes);
-
-export default router;
